data_extractor/code/new_ml_based_pipeline/new_ml_based_pipeline/requirements.txt
elastic-transport==7.16.0
elasticsearch==7.17.9
farm-haystack==1.23.0
fuzzywuzzy==0.18.0
pdf2image==1.17.0
pdfminer.six==20231228
PyMuPDF==1.23.18
pytesseract==0.3.10
